integrated/requirements.txt
PySide6>=6.5.0
opencv-python>=4.8.0
numpy>=1.24.0
PyOpenGL>=3.1.6

